+++ src/onegov.pdf/onegov/pdf/page_functions.py
@@ -4,6 +4,8 @@ from textwrap import wrap
 
 
 def empty_page_fn(cavnas, doc):
+    """ An empty header/footer. """
+
     pass
 
 
@@ -34,30 +36,36 @@ def page_fn_footer(canvas, doc):
 
 
 def page_fn_header(canvas, doc):
-    """ A standard header consisting of a title. The title is automatically
-    wrapped and shortened.
+    """ A standard header consisting of a title and the creation string. The
+    title is automatically wrapped and shortened.
 
     Example:
 
-        pdf = Pdf(file, author='OneGov')
+        pdf = Pdf(file, author='OneGov', crated='1.1.2000')
         pdf.init_a4_portrait(page_fn=draw_header)
 
     """
 
+    canvas.saveState()
+    canvas.setFont('Helvetica', 9)
     if doc.title:
-        canvas.saveState()
         lines = wrap(doc.title, 110)[:2]
         if len(lines) > 1:
             lines[1] = shorten(lines[1], 100)
         text = canvas.beginText()
-        text.setFont('Helvetica', 9)
         text.setTextOrigin(
             doc.leftMargin,
             doc.pagesize[1] - doc.topMargin * 2 / 3
         )
         text.textLines(lines)
         canvas.drawText(text)
-        canvas.restoreState()
+    if doc.created:
+        canvas.drawRightString(
+            doc.pagesize[0] - doc.rightMargin,
+            doc.pagesize[1] - doc.topMargin * 2 / 3,
+            doc.created
+        )
+    canvas.restoreState()
 
 
 def page_fn_header_and_footer(canvas, doc):
@@ -65,7 +73,7 @@ def page_fn_header_and_footer(canvas, doc):
 
     Example:
 
-        pdf = Pdf(file, author='OneGov')
+        pdf = Pdf(file, title='Title', created='1.1.2000', author='OneGov')
         pdf.init_a4_portrait(
             page_fn=draw_footer,
             page_fn_later=draw_header_and_footer